Make FloatValue class contain double instead of float

+++ source/float_value.cpp
@@ -8,7 +8,7 @@ FloatValue::FloatValue() {
 
 /// Constructor with parameters
 /// \param _value value to set
-FloatValue::FloatValue(float _value) {
+FloatValue::FloatValue(double _value) {
     value = _value;
 }
 
@@ -37,14 +37,14 @@ FloatValue::~FloatValue() = default;
 
 /// Value getter
 /// \return value
-float FloatValue::get() const {
+double FloatValue::get() const {
     return value;
 }
 
 
 /// Value setter
 /// \param _value the updated value of value
-void FloatValue::set(float _value) {
+void FloatValue::set(double _value) {
     value = _value;
 }
 
@@ -52,15 +52,15 @@ void FloatValue::set(float _value) {
 /// Addition operator for FloatValue
 /// \param other FloatValue object to add
 /// \return the sum of the values
-float FloatValue::operator+(const FloatValue &other) const {
+double FloatValue::operator+(const FloatValue &other) const {
     return value + other.value;
 }
 
 
-/// Addition operator for float
+/// Addition operator for double
 /// \param other integer to add
 /// \return the sum of value and other
-float FloatValue::operator+(float other) const {
+double FloatValue::operator+(double other) const {
     return value + other;
 }
 
@@ -68,15 +68,15 @@ float FloatValue::operator+(float other) const {
 /// Multiplication operator for FloatValue
 /// \param other FloatValue object to multiply
 /// \return the product of the values
-float FloatValue::operator*(const FloatValue &other) const {
+double FloatValue::operator*(const FloatValue &other) const {
     return value * other.value;
 }
 
 
-/// Multiplication operator for float
+/// Multiplication operator for double
 /// \param other integer to multiply
 /// \return the product of value and other
-float FloatValue::operator*(float other) const {
+double FloatValue::operator*(double other) const {
     return value * other;
 }
 
@@ -116,7 +116,7 @@ bool FloatValue::operator<(const FloatValue &other) const {
 /// Compare with equality operator
 /// \param other integer to compare to
 /// \return if the current value is equal to the other
-bool FloatValue::operator==(float other) const {
+bool FloatValue::operator==(double other) const {
     return std::fabs(value - other) < EPSILON;
 }
 
@@ -124,7 +124,7 @@ bool FloatValue::operator==(float other) const {
 /// Compare with inequality operator
 /// \param other integer to compare to
 /// \return if the current value is not equal to the other
-bool FloatValue::operator!=(float other) const {
+bool FloatValue::operator!=(double other) const {
     return std::fabs(value - other) >= EPSILON;
 }
 
@@ -132,7 +132,7 @@ bool FloatValue::operator!=(float other) const {
 /// Compare with greater than operator
 /// \param other integer to compare to
 /// \return if the current value is bigger than the other
-bool FloatValue::operator>(float other) const {
+bool FloatValue::operator>(double other) const {
     return value > other;
 }
 
@@ -140,6 +140,6 @@ bool FloatValue::operator>(float other) const {
 /// Compare with less than operator
 /// \param other integer to compare to
 /// \return if the current value is smaller than the other
-bool FloatValue::operator<(float other) const {
+bool FloatValue::operator<(double other) const {
     return value < other;
 }
